Python_Backend/requirements.txt
Keras_Preprocessing==1.1.0
Django==3.0.3
numpy==1.16.5
matplotlib==3.1.1
tensorflow==2.1.0
Pillow==7.0.0
